Add Database Service and Move Prisma Client To Database Service
